changed from add_action to add_filter
Just noticed these hooks should be filters and not actions. Tested, and working.

http://adambrown.info/p/wp_hooks/hook/pings_open
http://adambrown.info/p/wp_hooks/hook/comments_open

In a second thought, since we're using `__return_false()`, the class becomes a little less relevant. I would say it could be cleaner if the filters were outside the class. But, if you're looking for consistency then... better keep it this way :)